+++ www/go/core/views/extjs3/detail/Panel.js
@@ -63,6 +63,11 @@ go.detail.Panel = Ext.extend(Ext.Panel, {
 	},
 	
 	onChanges : function(entityStore, added, changed, destroyed) {
+
+		if(this.loading) {
+			return;
+		}
+
 		if(entityStore.entity.name === this.entityStore.entity.name) {
 			var entity = added[this.currentId] || changed[this.currentId] || false;
 
@@ -163,6 +168,8 @@ go.detail.Panel = Ext.extend(Ext.Panel, {
 		if(this.getTopToolbar()) {
 			this.getTopToolbar().setDisabled(false);
 		}
+
+		console.warn(data);
 		this.data = data;
 
 		var me = this;
@@ -186,6 +193,7 @@ go.detail.Panel = Ext.extend(Ext.Panel, {
 
 	load: function (id) {
 		this.currentId = id;
+		this.loading = true;
 		var me = this;
 		this.entityStore.single(id).then(function(entity) {
 			try {
@@ -197,7 +205,9 @@ go.detail.Panel = Ext.extend(Ext.Panel, {
 		}).catch(function(e) {
 			console.error(e);
 			Ext.MessageBox.alert(t("Not found"), "The requested page was not found");
-		});
+		}).finally(function() {
+			me.loading = false;
+		})
 	},
 
 	addCustomFields : function() {