+++ www/go/core/views/extjs3/util.js
@@ -1,3 +1,4 @@
+
 go.print = function(tmpl, data) {
 	var paper = document.getElementById('paper');
 
@@ -7,6 +8,9 @@ go.print = function(tmpl, data) {
 };
 
 go.util =  (function () {
+	var downloadFrame;
+
+
 	return {
 
 		clone: function(obj) {
@@ -299,7 +303,20 @@ go.util =  (function () {
 				if(inline) {
 					window.open(url);
 				} else {
-					document.location.href = url;
+					// document.location.href = url; //This causes connection errors with SSE or other simulanous XHR requests
+					// if(!downloadFrame) {
+					// 	downloadFrame = document.createElement('iframe');
+					// 	downloadFrame.id="downloader";
+					// 	downloadFrame.style.display = 'none';
+					// 	document.body.appendChild(downloadFrame);
+					// }
+					// downloadFrame.src = url;
+
+					if(!downloadFrame) {
+						var downloadFrame = document.createElement("a");
+					}
+					downloadFrame.href = url;
+					downloadFrame.click();
 				}
 			}
 		},