+++ www/views/Extjs3/javascript/form/HtmlEditor.js
@@ -193,12 +193,12 @@ Ext.extend(GO.form.HtmlEditor, Ext.form.HtmlEditor, {
 					console.warn(response);
 					var imgEl = null;
 					if (file.type.match(/^image\//)) {
-						domId = Ext.id(), img = '<img id="' + domId + '" src="' + go.Jmap.downloadUrl(response.blobId) + '" alt="' + file.name + '" />';
+						var domId = Ext.id(), img = '<img id="' + domId + '" src="' + go.Jmap.downloadUrl(response.blobId) + '" alt="' + file.name + '" />';
 						this.insertAtCursor(img);
 						imgEl = this.getDoc().getElementById(domId);
 					} 
 
-					this.fireEvent('attach', this, response.blobId, file, imgEl);
+					this.fireEvent('attach', this, response, file, imgEl);
 				}
 			});
 		}, this);